+++ src/commands/index.ts
@@ -11,6 +11,7 @@ import { setports } from './setports.js';
 import { queue } from './queue.js';
 import { allow } from './allow.js';
 import { diff } from './diff.js';
+import { voice } from './voice.js';
 
 export interface Command {
   data: SlashCommandBuilder;
@@ -30,4 +31,5 @@ commands.set(setports.data.name, setports as Command);
 commands.set(queue.data.name, queue);
 commands.set(allow.data.name, allow);
 commands.set(diff.data.name, diff);
+commands.set(voice.data.name, voice);